requirements.txt
Django==1.5
bpython
sockjs-tornado==1.0.0
tornado-redis==2.4.5
redis==2.8.0
django-extensions==1.1.0
dj-database-url==0.2.2
psycopg2==2.5.1
pymongo==2.6
uuid==1.30
